Contactors and Sleep update
Contactors will close when first 4200 received, open when CTRL C pressed.
BasicStatus in 4251 value changes based on data received in 8200 msg.
This is pretty klugey, the way that I have to change the BasicStatus...should be improved when time allows.
Contactors also a kludey as the will stay closed if the program crashes.
Buyer beware.